requirements.txt
SpeechRecognition
google-cloud-texttospeech
firebase-admin
numpy
opencv-python
gTTS
pygame
transformers
torch
Pillow
simpleaudio
pydub
